+++ packages/@expo/cli/CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add `-d` as an alias to `--dev-client`. ([#22925](https://github.com/expo/expo/pull/22925) by [@EvanBacon](https://github.com/EvanBacon))
 - Allow client-side device ids to reuse debugger sessions when restarting app. ([#22742](https://github.com/expo/expo/pull/22742) by [@byCedric](https://github.com/byCedric))
 - Enable inspector proxy with network support by default. ([#22936](https://github.com/expo/expo/pull/22936) by [@byCedric](https://github.com/byCedric))
+- Add `EXPO_OFFLINE` environment variable to disable network requests across the entire CLI.
 
 ### 🐛 Bug fixes